updates 15.05.2023
Upgraded email verification method because emails are sometimes late and this is the reason for random failed tests involving email verification. Also, we had an issue with a random failed deployment report step on GitHub Action due to the parallel execution, so I had to separate one large parallel execution into "smaller pieces". These current updates are demo versions and I will continue to follow this and find a better solution if it is needed.